Ejercicio terminado a falta de cambiar formulario a reactivo y página de detalle

+++ src/app/pages/admin/admin.component.ts
@@ -4,12 +4,10 @@ import { Ivacacion } from '../../vacacion';
 import { DatosVacacionService } from '../../services/datos-vacacion.service';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
-
-import { ChangeDetectionStrategy} from '@angular/core';
-import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { provideNativeDateAdapter } from '@angular/material/core';
+
 
 
 
@@ -21,14 +19,14 @@ import { provideNativeDateAdapter } from '@angular/material/core';
   templateUrl: './admin.component.html',
   styleUrl: './admin.component.sass'
 })
-  
+
 export class AdminComponent {
 
   viajes: Ivacacion[] = [];
   id: number;
   formularioOculto: boolean = true;
   viajeAEditar: Ivacacion = {
-    id: -1,
+    id: 0,
     viaje: "",
     fecha: "",
     imagen: "",
@@ -36,34 +34,44 @@ export class AdminComponent {
     detalle: ""
   }
 
-  constructor(private datosVacacionService: DatosVacacionService) {
-    this.viajes = datosVacacionService.actualizaViajes();
-  }
-
   displayedColumns: string[] = ['id', 'viaje', 'fecha', 'imagen', 'descripcion', 'editar', 'eliminar'];
   dataSource = this.viajes;
 
+  constructor(private datosVacacionService: DatosVacacionService) {
+
+    this.viajes = datosVacacionService.actualizaViajes();
+
+  }
+
   removeViaje(i: number) {
-    alert("¿Seguro que quieres elimninar este viaje?")
+
+    const confirmacion = this.confirmarBorrado();
+    if (!confirmacion) {
+      return;
+    }
     this.viajes = this.datosVacacionService.removeViaje(i)
   }
 
   editViaje(id: number): void {
-    this.formularioOculto = false;
 
-    console.log(`esto es lo que devuelve: ${id}`);
+    this.formularioOculto = false;
     this.viajeAEditar = this.datosVacacionService.getViajeById(id);
     this.viajeAEditar.id = id;
- 
-    console.log(`estos son los datos: ${this.viajeAEditar.viaje}`);
+
   }
 
   modificaViaje(viaje: Ivacacion) {
+
     this.viajeAEditar = viaje;
-    alert(this.viajeAEditar.viaje);
-    this.datosVacacionService.actualizaViajeById(this.viajeAEditar.id!, this.viajeAEditar);
+    this.datosVacacionService.modificaViajeById(this.viajeAEditar.id!, this.viajeAEditar);
     this.formularioOculto = true;
-    
+
+  }
+
+  confirmarBorrado(): boolean {
+
+    return window.confirm('¿Estás seguro de que deseas eliminar este viaje?');
+
   }
 
 }
